user and video bug fix

+++ hotube/src/controller/videoController.js
@@ -19,6 +19,10 @@ export const watch = async(req, res) => {
 export const getEdit = async(req, res) => {
   const id = req.params.id;
   const video = await Video.findById(id);
+  const userid = req.session.user._id;
+  if (String(video.owner) !== String(userid)) { // video.owner id 는 타입이 다른 그래서 타입도 맞춰줘야함.
+    return res.status(403).redirect("/");
+  };
   if (video === null) {
     return res.render("404", { pageTitle: "Video not found" })  
   }
@@ -26,20 +30,24 @@ export const getEdit = async(req, res) => {
 };
 
 export const postEdit = async(req, res) => {
-  const id = req.params.id;
-  const title = req.body.title;
-  const description = req.body.description;
-  const hashtags = req.body.hashtags;
-  const video = await Video.exists({ _id: id });
-  if (video === null) {
-    return res.status(404).render("404", { pageTitle: "Video not found" })  
+  const id = req.params.id
+  const userid = req.session.user._id;
+  const title = req.body.title
+  const description = req.body.description
+  const hashtags = req.body.hashtags
+  const video = await Video.find({ _id: id });
+  if (video._id === null) {
+    return res.status(404).render("404", { pageTitle: "Video not found" })
+  }
+  if (String(video.owner) !== String(userid)) {// video.owner id 는 타입이 다른 그래서 타입도 맞춰줘야함.
+    return res.status(403).redirect("/")
   }
   await Video.findByIdAndUpdate(id, {
     title: title,
     description: description,
     hashtags: Video.formatHashtags(hashtags),
   })
-  return res.redirect(`/videos/${id}`);
+  return res.redirect(`/videos/${id}`)
 };
 
 export const getUpload = (req, res) => {
@@ -70,7 +78,16 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
+  const userid = req.session.user._id;
   const id = req.params.id;
+  const video = await Video.findById(id);
+  if (!video) {
+      return res.status(404).render("404", { pageTitle: "Video not found." });
+  };
+  if (String(video.owner) !== String(userid)) {
+      // video.owner id 는 타입이 다른 그래서 타입도 맞춰줘야함.
+    return res.status(403).redirect("/");
+  };
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
